requirements.txt
tensorflow==2.6.0
keras==2.6.0
Pillow==7.1.2
matplotlib==3.2.2
numpy==1.19.5